+++ CHANGELOG.md
@@ -141,7 +141,7 @@
 
 ## 0.3.0 / 2019-09-25
 
-  * refactor: rename `collect.js` to `website-evidence-collector.js`
+  * refactor: rename `collect.js` to `website-evidence-collector.ts`
   * feature: make npm setup a symlink to have command `website-evidence-collector` available
   * bugfix: fix bug related to URL
   * documentation: add logo